% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uoc_psych_formats.R
\name{uoc_psych_pdf}
\alias{uoc_psych_pdf}
\title{University of Cologne Psychology Thesis}
\usage{
uoc_psych_pdf(
  fig_caption = TRUE,
  number_sections = FALSE,
  toc = TRUE,
  keep_tex = TRUE,
  md_extensions = NULL,
  includes = NULL,
  ...
)
}
\arguments{
\item{fig_caption}{\code{TRUE} to render figures with captions}

\item{number_sections}{\code{TRUE} to number section headings}

\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF.
Note that this argument does not control whether to keep the auxiliary
files (e.g., \file{.aux}) generated by LaTeX when compiling \file{.tex} to
\file{.pdf}. To keep these files, you may set \code{options(tinytex.clean =
FALSE)}.}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition of R Markdown. See the
\code{\link[rmarkdown]{rmarkdown_format}} for additional details.}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link[rmarkdown]{includes}} function).}

\item{...}{Further arguments to pass to
\code{\link[bookdown]{pdf_document2}} or
\code{\link[bookdown]{word_document2}}.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{rmarkdown::render()}}.
}
\description{
Template for theses in psychology at the University of Cologne in PDF format.
}
\details{
When creating PDF documents the YAML option \code{classoption} is passed
to the class options of the LaTeX apa6 document class. In this case,
additional options are available. Refer to the \code{apa6} document class
\href{ftp://ftp.fu-berlin.de/tex/CTAN/macros/latex/contrib/apa6/apa6.pdf}{documentation}
to find out about class options such as paper size.

Please refer to the \href{https://frederikaust.com/papaja_man/r-markdown-components.html#yaml-front-matter}{\pkg{papaja} online-manual}
for additional information on available YAML front matter settings.

When creating PDF documents the output device for figures defaults to
\code{c("pdf", "png")}, so that each figure is saved in all four formats
at a resolution of 300 dpi.
}
\seealso{
\code{\link[bookdown:html_document2]{bookdown::pdf_document2()}}, \code{\link[bookdown:html_document2]{bookdown::word_document2()}}
}
